services/earnings/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
